API: move product stats to core
